Move keyToAddress into derivation module
