+++ packages/iov-ethereum/src/derivation.ts
@@ -1,3 +1,4 @@
+import { Algorithm, PublicKeyBundle } from "@iov/base-types";
 import { Address } from "@iov/bcp-types";
 import { Keccak256 } from "@iov/crypto";
 import { Encoding } from "@iov/encoding";
@@ -20,4 +21,25 @@ export function isValidAddress(address: string): boolean {
     return true;
   }
   return false;
+}
+
+export function toChecksumAddress(address: string): Address {
+  // https://github.com/ethereum/EIPs/blob/master/EIPS/eip-55.md
+  const addressLower = address.toLowerCase().replace("0x", "") as Address;
+  const addressHash = toHex(new Keccak256(toAscii(addressLower)).digest());
+  let checksumAddress = "0x";
+  for (let i = 0; i < 40; i++) {
+    checksumAddress += parseInt(addressHash[i], 16) > 7 ? addressLower[i].toUpperCase() : addressLower[i];
+  }
+  return checksumAddress as Address;
+}
+
+export function keyToAddress(pubkey: PublicKeyBundle): Address {
+  if (pubkey.algo !== Algorithm.Secp256k1 || pubkey.data.length !== 65 || pubkey.data[0] !== 0x04) {
+    throw new Error(`Invalid pubkey data input: ${pubkey}`);
+  }
+  const hash = toHex(new Keccak256(pubkey.data.slice(1)).digest());
+  const lastFortyChars = hash.slice(-40);
+  const addressString = toChecksumAddress("0x" + lastFortyChars);
+  return addressString as Address;
 }
