+++ packages/iov-ethereum/src/ethereumcodec.ts
@@ -1,6 +1,5 @@
-import { Algorithm, ChainId, PostableBytes, PublicKeyBundle } from "@iov/base-types";
+import { ChainId, PostableBytes } from "@iov/base-types";
 import {
-  Address,
   Nonce,
   PrehashType,
   SignableBytes,
@@ -11,26 +10,15 @@ import {
   TxCodec,
   UnsignedTransaction,
 } from "@iov/bcp-types";
-import { ExtendedSecp256k1Signature, Keccak256 } from "@iov/crypto";
+import { ExtendedSecp256k1Signature } from "@iov/crypto";
 import { Encoding } from "@iov/encoding";
 
-import { isValidAddress } from "./derivation";
+import { isValidAddress, keyToAddress } from "./derivation";
 import { BlknumForkState, Eip155ChainId, eip155V, toRlp } from "./encoding";
 import { Serialization } from "./serialization";
 import { encodeQuantity, encodeQuantityString, hexPadToEven } from "./utils";
 
-const { fromHex, toAscii, toHex } = Encoding;
-
-export function toChecksumAddress(address: string): Address {
-  // https://github.com/ethereum/EIPs/blob/master/EIPS/eip-55.md
-  const addressLower = address.toLowerCase().replace("0x", "") as Address;
-  const addressHash = toHex(new Keccak256(toAscii(addressLower)).digest());
-  let checksumAddress = "0x";
-  for (let i = 0; i < 40; i++) {
-    checksumAddress += parseInt(addressHash[i], 16) > 7 ? addressLower[i].toUpperCase() : addressLower[i];
-  }
-  return checksumAddress as Address;
-}
+const { fromHex } = Encoding;
 
 export const ethereumCodec: TxCodec = {
   bytesToSign: (unsigned: UnsignedTransaction, nonce: Nonce): SigningJob => {
@@ -96,14 +84,6 @@ export const ethereumCodec: TxCodec = {
   parseBytes: (bytes: PostableBytes, chainId: ChainId): SignedTransaction => {
     throw new Error(`Not implemented bytes: ${bytes}, chainId: ${chainId}`);
   },
-  keyToAddress: (pubkey: PublicKeyBundle): Address => {
-    if (pubkey.algo !== Algorithm.Secp256k1 || pubkey.data.length !== 65 || pubkey.data[0] !== 0x04) {
-      throw new Error(`Invalid pubkey data input: ${pubkey}`);
-    }
-    const hash = toHex(new Keccak256(pubkey.data.slice(1)).digest());
-    const lastFortyChars = hash.slice(-40);
-    const addressString = toChecksumAddress("0x" + lastFortyChars);
-    return addressString as Address;
-  },
+  keyToAddress: keyToAddress,
   isValidAddress: isValidAddress,
 };
